Implement option to automatically load symbol maps
The loader will scan the directory from which the binary is loaded for a map file with the same name as the binary file.

If it fails to find a symbol map it will revert to asking the user if they want to load one manually.

The option is enabled by default.